src/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiomysql>=0.2.0",
    "aioodbc>=0.5.0",
    "annotated-types>=0.7.0",
    "anyio>=4.7.0",
    "apscheduler>=3.10.4",
    "asttokens>=3.0.0",
    "cached-property>=1.5.2",
    "certifi>=2024.12.14",
    "cffi>=1.17.1",
    "charset-normalizer>=3.4.1",
    "click>=8.1.8",
    "colorama>=0.4.6",
    "cryptography>=44.0.0",
    "defusedxml>=0.7.1",
    "dnspython>=2.7.0",
    "ecdsa>=0.19.0",
    "exchangelib>=5.4.3",
    "executing>=2.1.0",
    "fastapi[standard]>=0.115.6",
    "future>=1.0.0",
    "greenlet>=3.1.1",
    "gevent>=24.11.1",
    "h11>=0.14.0",
    "icecream>=2.1.3",
    "idna>=3.10",
    "iniconfig>=2.0.0",
    "isodate>=0.7.2",
    "jinja2>=3.1.5",
    "ldap3>=2.9.1",
    "lxml>=5.3.0",
    "markupsafe>=3.0.2",
    "oauthlib>=3.2.2",
    "packaging>=24.2",
    "pluggy>=1.5.0",
    "pyasn1>=0.6.1",
    "pycparser>=2.22",
    "pycryptodome>=3.21.0",
    "pydantic>=2.10.6",
    "pydantic-core>=2.27.2",
    "pygments>=2.18.0",
    "pymysql>=1.1.1",
    "pyodbc>=5.3.0",
    "pyspnego>=0.11.2",
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "python-dotenv>=1.0.1",
    "python-jose>=3.3.0",
    "python-tds>=1.15.0",
    "pytz>=2024.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "requests-ntlm>=1.3.0",
    "requests-oauthlib>=2.0.0",
    "rsa>=4.9",
    "setuptools>=75.8.0",
    "shellescape>=3.8.1",
    "six>=1.17.0",
    "sniffio>=1.3.1",
    "sqlalchemy>=2.0.36",
    "sqlmodel>=0.0.27",
    "starlette>=0.41.3",
    "typing-extensions>=4.12.2",
    "tzdata>=2024.2",
    "tzlocal>=5.2",
    "urllib3>=2.3.0",
    "uvicorn>=0.34.0",
    "gunicorn>=23.0.0",
    "slowapi>=0.1.9",
    "pydantic-settings>=2.12.0",
    "passlib>=1.7.4",
    "celery[redis]>=5.5.0",
    "alembic>=1.14.0",
    "redis>=5.3.1",
    "opentelemetry-api>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-sdk>=1.29.0",
    "prometheus-client>=0.21.1",
    "psutil>=6.1.1",
    "kombu>=5.5.0",
    "billiard>=4.2.4",
]

[dependency-groups]
dev = [
    "ruff>=0.9.1",
]

